% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addSizeClassToCatch}
\alias{addSizeClassToCatch}
\title{addSizeClassToCatch}
\usage{
addSizeClassToCatch(basket, catch)
}
\arguments{
\item{basket}{A data frame containing the basket information from Andes}

\item{catch}{A data frame containing the catch information from Andes}
}
\value{
the merged catch card
}
\description{
This function fills in the catch with the missing size class 2 entries.   
Andes does not provide a way to add Size class directly, Size classes are entered via the basket form.
}
\seealso{
Other general_use: 
\code{\link{formatForOracle}()},
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{toOracle}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimens}()},
\code{\link{tweakUniversal}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
