% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSQLScript.R
\docType{data}
\name{base}
\alias{base}
\title{script}
\format{
An object of class \code{character} of length 1.
}
\usage{
base
}
\value{
nothing
}
\description{
This scripts generates the SQL queries needed to populate ANdes DB with data coming from ESE
}
\seealso{
Other internal: 
\code{\link{cleanEse}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertLengthUnits}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{convertYesNo}()},
\code{\link{createInsertIntoObsFields}()},
\code{\link{createInsertIntoObsGrp}()},
\code{\link{createInsertIntoSampReq}()},
\code{\link{getEseTables}()},
\code{\link{getKeyFields}()},
\code{\link{getTblKey}()},
\code{\link{get_value}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
\keyword{datasets}
