% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{basketTweaks}
\alias{basketTweaks}
\title{basketTweaks}
\usage{
basketTweaks(x = NULL, quiet = FALSE)
}
\arguments{
\item{x}{default is \code{NULL}.  Basket data frame}

\item{quiet}{default is \code{FALSE} Determines whether or not the script should output 
informational messages}
}
\value{
a basket data frame
}
\description{
This function will perform tweaks to data coming from ANdes before 
it is imported to Oracle.
}
\seealso{
Other general_use: 
\code{\link{addSizeClassToCatch}()},
\code{\link{catchTweaks}()},
\code{\link{eseExtractor}()},
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setTweaks}()},
\code{\link{specimenTweaks}()},
\code{\link{toOracle}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
