% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{charToBinary}
\alias{charToBinary}
\title{charToBinary}
\usage{
charToBinary(x = NULL, bool = T)
}
\arguments{
\item{x}{default is \code{NULL}.  This is a vector of values.  Case-insensitive, and anything
other than "true" or "false" will become NA}

\item{bool}{default is \code{TRUE}. By default, this function turns values into TRUE/FALSE, but 
if \code{bool} is F, it will output "Y" and "N" (for "yes" and "no")}
}
\description{
This function converts the values of "True" and "False" output by andes into either
TRUE/FALSE or Y/N
}
\seealso{
Other internal: 
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{keepFieldsBaskets}()},
\code{\link{keepFieldsCatches}()},
\code{\link{keepFieldsMissions}()},
\code{\link{keepFieldsSets}()},
\code{\link{keepFieldsSpecimens}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()},
\code{\link{redistributeMixedCatch}()},
\code{\link{transmogrifyBaskets}()},
\code{\link{transmogrifyCatches}()},
\code{\link{transmogrifyLV1_OBS}()},
\code{\link{transmogrifyMissions}()},
\code{\link{transmogrifySets}()},
\code{\link{transmogrifySpecimens}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
