% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkPrimaryFields.R
\name{checkPrimaryFields}
\alias{checkPrimaryFields}
\title{checkPrimaryFields}
\usage{
checkPrimaryFields(primaryTable = NULL)
}
\arguments{
\item{primaryTable}{default is \code{NULL}.  This is a connection object from}
}
\value{
nothing - just outputs messages
}
\description{
This function verifies that key fields within one ESE table exist in the other 
tables in the ESE.  For example, if a SETNO exists in ESE_CATCHES that does not exist in ESE_SETS,
a message would be output.
}
\seealso{
Other general_use: 
\code{\link{addSizeClassToCatch}()},
\code{\link{eseExtractor}()},
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{toOracle}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
