% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanEse}
\alias{cleanEse}
\title{cleanEse}
\usage{
cleanEse()
}
\value{
nothing - just remove data frames from environment
}
\description{
This function removes any ESE objects from the R environment
}
\seealso{
Other internal: 
\code{\link{TF2YN}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
