% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanfields}
\alias{cleanfields}
\title{cleanfields}
\usage{
cleanfields(data = NULL)
}
\arguments{
\item{data}{default is \code{NULL}.  This is a field in a data frame}
}
\description{
This function takes a field and replaces various odd "smart" curly quotes and 
apostrophes and replaces them with normal versions
}
\seealso{
Other internal: 
\code{\link{TF2YN}()},
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
