% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanStrata}
\alias{cleanStrata}
\title{cleanStrata}
\usage{
cleanStrata(x)
}
\description{
This function takes the strata field, and ensures only the first 3 characters are 
returned#' @param x default is \code{NULL}.  This is a field in a data frame
}
\seealso{
Other internal: 
\code{\link{TF2YN}()},
\code{\link{cleanEse}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
