% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{colTypeConverter}
\alias{colTypeConverter}
\title{colTypeConverter}
\usage{
colTypeConverter(df = NULL)
}
\arguments{
\item{df}{default is \code{NULL} the dataframe that needs column types changed}
}
\value{
the same df, but with attributes
}
\description{
This function takes a dataframe and tries to convert the various columns into the 
appropriate type according to their contents
}
\seealso{
Other general_use: 
\code{\link{loadData}()},
\code{\link{loadESEData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{tweakBasketsPostProcessing}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimensRaw}()},
\code{\link{tweakUniversal}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
