% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{convert_to_DDMM}
\alias{convert_to_DDMM}
\title{convert_to_DDMM}
\usage{
convert_to_DDMM(decimal_degrees, roundDec = NULL)
}
\arguments{
\item{decimal_degrees}{default is \code{NULL}. This is the value you wish to convert}

\item{roundDec}{default is \code{NULL}.  This is how many decimals you would like your final 
output to have.  If you do not want to round the result, you can leave it as NULL}
}
\value{
la coordinate in DDMM.MM
}
\description{
This function converts coordinates from decimal degrees to DDMM.MM
}
\author{
Mike McMahon, \email{mike.mcmahon@dfo-mpo.gc.ca}
}
