% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUtilities.R
\name{convertHOWOBT}
\alias{convertHOWOBT}
\title{convertHOWOBT}
\usage{
convertHOWOBT(x)
}
\arguments{
\item{x}{default is \code{NULL}. This is the string to be processed.}
}
\value{
converted FORCE codes
}
\description{
This function returns converted obtained_code codes
}
\seealso{
Other internal: 
\code{\link{TF2YN}()},
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
