% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.utilities.R
\name{convertLengthUnits}
\alias{convertLengthUnits}
\title{convertLengthUnits}
\usage{
convertLengthUnits(x)
}
\arguments{
\item{x}{default is \code{NULL}.}
}
\value{
numeric value representing of index
}
\description{
This function returns index values used by andes for given string in ESE
}
\seealso{
Other internal: 
\code{\link{cleanEse}()},
\code{\link{cleanfields}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertSex}()},
\code{\link{convertStrata}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{convertYesNo}()},
\code{\link{getEseTables}()},
\code{\link{getKeyFields}()},
\code{\link{getTblKey}()},
\code{\link{get_value}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Pablo Vergara, \email{pablo.vergara@dfo-mpo.gc.ca}
}
\concept{internal}
