% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.utilities.R
\name{convertSex}
\alias{convertSex}
\title{convertsex}
\usage{
convertSex(x)
}
\value{
converted sex codes
}
\description{
This function returns converted sex codes
}
\seealso{
Other internal: 
\code{\link{base}},
\code{\link{cleanEse}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertLengthUnits}()},
\code{\link{convertStrata}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{convertYesNo}()},
\code{\link{createInsertIntoObsFields}()},
\code{\link{createInsertIntoObsGrp}()},
\code{\link{createInsertIntoSampReq}()},
\code{\link{createInsertIntoSex}()},
\code{\link{createInsertIntoStrata}()},
\code{\link{getEseTables}()},
\code{\link{getKeyFields}()},
\code{\link{getTblKey}()},
\code{\link{get_value}()}
}
\author{
Pablo Vergara, \email{pablo.vergara@dfo-mpo.gc.ca}
}
\concept{internal}
