% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.utilities.R
\name{convertYesNo}
\alias{convertYesNo}
\title{convertYesNo}
\usage{
convertYesNo(x)
}
\arguments{
\item{x}{default is \code{NULL}. This is the string to be processed.}
}
\value{
converted characters
}
\description{
This function returns converted "Y" or "N' to 0 or 1
}
\seealso{
Other internal: 
\code{\link{cleanEse}()},
\code{\link{cleanfields}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertLengthUnits}()},
\code{\link{convertSex}()},
\code{\link{convertStrata}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{getEseTables}()},
\code{\link{getKeyFields}()},
\code{\link{getTblKey}()},
\code{\link{get_value}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Pablo Vergara, \email{pablo.vergara@dfo-mpo.gc.ca}
}
\concept{internal}
