% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEseToAndesSQL.R
\name{createInsertIntoStrata}
\alias{createInsertIntoStrata}
\title{createInsertIntoStrata}
\usage{
createInsertIntoStrata(protocol, rawFile, stratumFile)
}
\value{
SQL queries
}
\description{
This function returns the SQL queries for the shared_models_stratum in ANDES using data coming from ESE
}
\seealso{
Other internal: 
\code{\link{base}},
\code{\link{cleanEse}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertLengthUnits}()},
\code{\link{convertSex}()},
\code{\link{convertStrata}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{convertYesNo}()},
\code{\link{createInsertIntoObsFields}()},
\code{\link{createInsertIntoObsGrp}()},
\code{\link{createInsertIntoSampReq}()},
\code{\link{createInsertIntoSex}()},
\code{\link{getEseTables}()},
\code{\link{getKeyFields}()},
\code{\link{getTblKey}()},
\code{\link{get_value}()}
}
\author{
Pablo Vergara, \email{pablo.vergara@dfo-mpo.gc.ca}
}
\concept{internal}
