% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exeExtractor.R
\name{eseExtractor}
\alias{eseExtractor}
\title{eseExtractor}
\usage{
eseExtractor(
  cxnObj = NULL,
  mission = NULL,
  tabs = NULL,
  quiet = FALSE,
  env = .GlobalEnv
)
}
\arguments{
\item{cxnObj}{default is \code{NULL}.  This is a connection object from 
\code{Mar.utils::make_oracle_cxn()}.}

\item{mission}{default is \code{NULL}.  This is a vector of one or more mission identifiers 
(e.g "CAR2021240") that will be used to limit the extractions to particular mission(s). 
\code{"ALL"} is also valid, and will return all records for all of the specified tables. 
Use with caution.}

\item{tabs}{default is \code{NULL}.  This is a vector the ESE tables that should be 
extracted.  By default, all will be extracted, but fewer names can be sent to extract a subset.  
Valid values include \code{c("ESE_MISSIONS", "ESE_SETS", "ESE_CATCHES", "ESE_BASKETS", 
"ESE_SPECIMENS", "ESE_LV1_OBSERVATIONS")}.}

\item{quiet}{default is \code{FALSE} Determines whether or not the script should output 
informational messages}

\item{env}{default is \code{.GlobalEnv} The environment where the data should be loaded}
}
\value{
nothing - just loads data
}
\description{
This function extracts all of the ESE data for a specified MISSION from the 
GROUNDFISH schema.
}
\seealso{
Other general_use: 
\code{\link{addSizeClassToCatch}()},
\code{\link{basketTweaks}()},
\code{\link{catchTweaks}()},
\code{\link{checkPrimaryFields}()},
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setTweaks}()},
\code{\link{specimenTweaks}()},
\code{\link{toOracle}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
