% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatForROracle.R
\name{formatForOracle}
\alias{formatForOracle}
\title{formatForROracle}
\usage{
formatForOracle(source_df = NULL)
}
\arguments{
\item{source_df}{default is \code{NULL}.This is a data frame containing the data that should be 
loaded to Oracle.It should correspond with the structure used by one of ESE_MISSIONS, ESE_SETS, 
ESE_CATCHES, ESE_BASKETS, ESE_SPECIMENS, ESE_LV1_OBSERVATIONS}
}
\value{
data frame, but with attr()
}
\description{
This function applies field-specific oracle formatting to the ESE objects so that 
they each get the appropriate field types when they are loaded (e.g. numeric, date, correct # of 
decimal places, etc)
}
\note{
This may
}
\seealso{
Other general_use: 
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{toOracle}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimens}()},
\code{\link{tweakUniversal}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
