% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{GEORGES_5Z}
\alias{GEORGES_5Z}
\title{GEORGES_5Z Metadata}
\format{
data.frame
\describe{
\item{STRATUM}{survey stratum code}
\item{PRIMARY}{integer indicating required number of primary stations}
\item{ALTERNATE}{integer indicating required number of alternate stations}
\item{SECONDARY}{integer indicating required number of secondary stations}
}
}
\usage{
GEORGES_5Z
}
\description{
This data.frame includes all of the strata sampled during modern Georges 5Z surveys
and the required number of primary, secondary and alternate stations for each.
}
\keyword{datasets}
