% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUtilities.R
\name{get_value}
\alias{get_value}
\title{get_value}
\usage{
get_value(myKey, mylookupvector)
}
\arguments{
\item{myKey}{default is \code{NULL}.}

\item{mylookupvector}{default is \code{NULL}.}
}
\value{
lookup value
}
\description{
This function returns the  lookup value in a vector
}
\seealso{
Other internal: 
\code{\link{TF2YN}()},
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Pablo Vergara, \email{pablo.vergara@dfo-mpo.gc.ca}
}
\concept{internal}
