% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getEseTables}
\alias{getEseTables}
\title{getEseTables}
\usage{
getEseTables()
}
\value{
character vector of the 6 core ESE tables
}
\description{
This function returns the names of the ese tables.
}
\seealso{
Other internal: 
\code{\link{cleanEse}()},
\code{\link{cleanfields}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertLengthUnits}()},
\code{\link{convertSex}()},
\code{\link{convertStrata}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{convertYesNo}()},
\code{\link{getKeyFields}()},
\code{\link{getTblKey}()},
\code{\link{get_value}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
