% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getTblKey}
\alias{getTblKey}
\title{getTblKey}
\usage{
getTblKey(df = NULL, keyFields = NULL)
}
\arguments{
\item{df}{default is \code{NULL}.  This is a connection object from Mar.utils::make_oracle_cxn().}

\item{keyFields}{default is \code{NULL}.  This is a connection object from Mar.utils::make_oracle_cxn().}
}
\value{
nothing - just loads data to environment
}
\description{
This function takes a data frame and a vector of field names from that data frame.
It concatenates together all of the combinations of values from those key fields into a vector.  
When appropriate fields are chosen as \code{keyFields}, the result can be used to create a 
primary key for identifying unique rows
}
\seealso{
Other internal: 
\code{\link{base}},
\code{\link{cleanEse}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertLengthUnits}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{convertYesNo}()},
\code{\link{createInsertIntoObsFields}()},
\code{\link{createInsertIntoObsGrp}()},
\code{\link{createInsertIntoSampReq}()},
\code{\link{getEseTables}()},
\code{\link{getKeyFields}()},
\code{\link{get_value}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
