% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inserter.R
\name{inserter}
\alias{inserter}
\title{inserter}
\usage{
inserter(
  cxnObj = NULL,
  source_df = NULL,
  target_table = NULL,
  target_schema = NULL
)
}
\arguments{
\item{cxnObj}{default is \code{NULL}.  This is a connection object from make_oracle_cxn().}

\item{source_df}{default is \code{NULL}.  This is a df containing the data that should be loaded 
to Oracle}

\item{target_table}{default is \code{NULL}.  This is the name of the specific Oracle table within
\code{target_schema} where the data should be loaded}

\item{target_schema}{default is \code{NULL}.  This is the name of the Oracle schema where the 
data stored in\code{source_df} should be loaded}
}
\value{

}
\description{
This function loads all of the csv files in a specified folder into the global 
environment.  The object names will be identical to the original file names.
}
\seealso{
Other general_use: 
\code{\link{loadData}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
