% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepFields.R
\name{keepFieldsSets}
\alias{keepFieldsSets}
\title{keepFieldsSets}
\usage{
keepFieldsSets(df = NULL, mission = NULL)
}
\arguments{
\item{df}{default is \code{NULL}.  This is the andes set_data data frame}

\item{mission}{default is \code{NULL}. This is the cleaned mission id (e.g. CAR2022102)}
}
\description{
This function subsets the andes sets_data, retaining only the fields usable by the 
ESE_SETS table
}
\seealso{
Other internal: 
\code{\link{applySubsampling}()},
\code{\link{charToBinary}()},
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{keepFieldsBaskets}()},
\code{\link{keepFieldsCatches}()},
\code{\link{keepFieldsMissions}()},
\code{\link{keepFieldsSpecimens}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()},
\code{\link{transmogrifyBaskets}()},
\code{\link{transmogrifyCatches}()},
\code{\link{transmogrifyLV1_OBS}()},
\code{\link{transmogrifyMissions}()},
\code{\link{transmogrifySets}()},
\code{\link{transmogrifySpecimens}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
