% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keepFields.R
\name{keepFieldsSpecimens}
\alias{keepFieldsSpecimens}
\title{keepFieldsSpecimens}
\usage{
keepFieldsSpecimens(df = NULL, mission = NULL)
}
\arguments{
\item{df}{default is \code{NULL}.  This is the andes specimen_data data frame}

\item{mission}{default is \code{NULL}. This is the cleaned mission id (e.g. CAR2022102)}
}
\description{
This function subsets the andes specimen_data, retaining only the fields usable by 
both the ESE_SPECIMENS and ESE_LV1_OBSERVATIONS table
}
\seealso{
Other internal: 
\code{\link{charToBinary}()},
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{keepFieldsBaskets}()},
\code{\link{keepFieldsCatches}()},
\code{\link{keepFieldsMissions}()},
\code{\link{keepFieldsSets}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()},
\code{\link{redistributeMixedCatch}()},
\code{\link{transmogrifyBaskets}()},
\code{\link{transmogrifyCatches}()},
\code{\link{transmogrifyLV1_OBS}()},
\code{\link{transmogrifyMissions}()},
\code{\link{transmogrifySets}()},
\code{\link{transmogrifySpecimens}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
