% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{loadData}
\alias{loadData}
\title{loadData}
\usage{
loadData(dataPath = NULL)
}
\arguments{
\item{dataPath}{default is \code{NULL} the folder containing files to be loaded into R.  If left 
as NULL, csv files in the\code{sampleAndesData} folder will be loaded}
}
\value{
nothing - just loads data to environment
}
\description{
This function loads all of the csv files in a specified folder into the global 
environment.  The object names will be identical to the original file names.
}
\seealso{
Other general_use: 
\code{\link{addSizeClassToCatch}()},
\code{\link{basketTweaks}()},
\code{\link{catchTweaks}()},
\code{\link{formatForOracle}()},
\code{\link{lv1Tweaks}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{setTweaks}()},
\code{\link{specimenTweaks}()},
\code{\link{toOracle}()},
\code{\link{universalTweaks}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
