% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadESEData.R
\name{loadESEData}
\alias{loadESEData}
\title{loadESEData}
\usage{
loadESEData(
  cxnObj = NULL,
  source_df = NULL,
  target_table = NULL,
  confirmOverwrite = F
)
}
\arguments{
\item{cxnObj}{default is \code{NULL}.  This is a connection object from 
Mar.utils::make_oracle_cxn(). It MUST use ROracle - RODBC is not supported.  The account 
specified by this needs permission to write tables in the \code{target_schema}.}

\item{source_df}{default is \code{NULL}.  This is a data frame containing the data that should be 
loaded to Oracle.  It should correspond with the structure used by one of ESE_MISSIONS, ESE_SETS, 
ESE_CATCHES, ESE_BASKETS, ESE_SPECIMENS, ESE_LV1_OBSERVATIONS}

\item{target_table}{default is \code{NULL}.  This is the name of the specific Oracle table within 
the user schema (i.e. GROUNDFISH) where the data should go}

\item{confirmOverwrite}{default is \code{F}. This must be set to \code{T} to append data to the 
production ESE tables}
}
\value{
nothing, but data gets loaded to the selected location in Oracle.
}
\description{
This function loads any of 6 known data frame objects into Oracle.  If an unknown 
\code{target_table} is specified, that table will be created (and deleted, if it already existed)
If the \code{target_table} is an existing ESE_* table, and \code{confirmOverwrite} is TRUE, then 
the new data will be appended.
}
\seealso{
Other general_use: 
\code{\link{colTypeConverter}()},
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{tweakBasketsPostProcessing}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimensRaw}()},
\code{\link{tweakUniversal}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
