% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_oracle_cxn.r
\name{make_oracle_cxn}
\alias{make_oracle_cxn}
\title{make_oracle_cxn}
\usage{
make_oracle_cxn(
  usepkg = "roracle",
  fn.oracle.username = "_none_",
  fn.oracle.password = "_none_",
  fn.oracle.dsn = "_none_",
  quietly = FALSE
)
}
\arguments{
\item{usepkg}{default is \code{'roracle'}. This indicates whether the connection to Oracle should
use \code{'rodbc'} or \code{'roracle'} to connect.  rodbc is slightly easier to setup, but
roracle will extract data ~ 5x faster.}

\item{fn.oracle.username}{default is \code{'_none_'} This is your username for
accessing oracle objects. If you have a value for \code{oracle.username} 
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.password}{default is \code{'_none_'} This is your password for
accessing oracle objects. If you have a value for \code{oracle.password}  
stored in your environment (e.g. from an rprofile file), this can be left out
and that value will be used.  If a value for this is provided, it will take 
priority over your existing value.}

\item{fn.oracle.dsn}{default is \code{'_none_'} This is your dsn/ODBC
identifier for accessing oracle objects. If you have a value for 
\code{oracle.dsn} stored in your environment (e.g. from an rprofile file), 
this can be left and that value will be used.  If a value for this is 
provided, it will take priority over your existing value.}

\item{quietly}{default is \code{FALSE}  This indicates whether or not status messages should be 
shown.}
}
\value{
a list with 3 elements: 
\itemize{
\item \code{usepkg} - this just stores which of 'roracle' or 'rodbc' was selected
\item \code{channel} - this is the actual connection object
\item \code{thecmd} - this is the command you need to select data using the selected connection 
type - i.e. "dbGetQuery" for ROracle or "sqlQuery" for RODBC
}
}
\description{
This function facilitates creating a connection to Oracle, and
allows connection via RODBC or ROracle, depending on the value of usepkg.
}
\note{
This function was stolen directly from https://github.com/Maritimes/Mar.utils, but I can 
live with it, since I wrote it. :)
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{utilities}
