% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{populate_DATA_DESC}
\alias{populate_DATA_DESC}
\title{populate_DATA_DESC}
\usage{
populate_DATA_DESC(x)
}
\arguments{
\item{x}{default is \code{NULL}.  This is a field in a data frame}
}
\description{
This function populates the DATA_DESCR field of the LV1_OBSERVATION DATA_DESC 
appropriately depending on the values found
}
\seealso{
Other internal: 
\code{\link{charToBinary}()},
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{keepFieldsBaskets}()},
\code{\link{keepFieldsCatches}()},
\code{\link{keepFieldsMissions}()},
\code{\link{keepFieldsSets}()},
\code{\link{keepFieldsSpecimens}()},
\code{\link{meters2Fathoms}()},
\code{\link{reFormatSpecimen}()},
\code{\link{redistributeMixedCatch}()},
\code{\link{transmogrifyBaskets}()},
\code{\link{transmogrifyCatches}()},
\code{\link{transmogrifyLV1_OBS}()},
\code{\link{transmogrifyMissions}()},
\code{\link{transmogrifySets}()},
\code{\link{transmogrifySpecimens}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
