% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reDistributeMixedCatch}
\alias{reDistributeMixedCatch}
\title{reDistributeMixedCatch}
\usage{
reDistributeMixedCatch(catch = NULL, basket = NULL)
}
\arguments{
\item{catch}{default is \code{NULL}.}

\item{basket}{default is \code{NULL}.}
}
\description{
This function does stuff
}
\seealso{
Other internal: 
\code{\link{charToBinary}()},
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{keepFieldsBaskets}()},
\code{\link{keepFieldsCatches}()},
\code{\link{keepFieldsMissions}()},
\code{\link{keepFieldsSets}()},
\code{\link{keepFieldsSpecimens}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()},
\code{\link{transmogrifyBaskets}()},
\code{\link{transmogrifyCatches}()},
\code{\link{transmogrifyLV1_OBS}()},
\code{\link{transmogrifyMissions}()},
\code{\link{transmogrifySets}()},
\code{\link{transmogrifySpecimens}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
