% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reFormatSpecimen}
\alias{reFormatSpecimen}
\title{reFormatSpecimen}
\usage{
reFormatSpecimen(x = NULL)
}
\arguments{
\item{x}{default is \code{NULL}.  This is a field in a data frame}
}
\description{
This function takes the tempSpecimen object, and populates the DATA_DEC 
appropriately depending on the values found
}
\seealso{
Other internal: 
\code{\link{TF2YN}()},
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
