% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{setExperimentType}
\alias{setExperimentType}
\title{setExperimentType}
\usage{
setExperimentType(x)
}
\arguments{
\item{x}{set object from ANdes raw data}
}
\value{
The experiment code
}
\description{
This function sets the proper experiment type to match existing data structures in Maritime and Gulf region
. Based on a mix of experiment typer and set results
}
\seealso{
Other general_use: 
\code{\link{formatForOracle}()},
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{toOracle}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimens}()},
\code{\link{tweakUniversal}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
