% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSelect.R
\name{setSelect}
\alias{setSelect}
\title{setSelect}
\usage{
setSelect(
  MaritimesSurvey = NULL,
  stationData = NULL,
  stationData_StratField = NULL,
  strataShp = NULL,
  strataShp_StratField = NULL,
  avoidShp = NULL,
  localCRS = NULL,
  minDistNM = 4,
  addNafoInfo = T,
  oceansAreas = NULL,
  tryXTimes = 100
)
}
\arguments{
\item{MaritimesSurvey}{default is \code{NULL}.  By default, the following parameters must be 
supplied for this function to work:
\itemize{
  \item \code{stationData}
  \item \code{stationData_StratField}
  \item \code{strataShp}
  \item \code{strataShp_StratField}
  \item \code{localCRS}
}
However, if \code{MaritimesSurvey} is set to one of the Maritimes-specific identifiers below, the
fields above will be auto-populated:
\itemize{
 \item \code{"4VW_Spring"}
 \item \code{"4X_Spring"}
 \item \code{"4VWX_Summer"}
 \item \code{"5Z_Georges"}
 \item \code{"custom"}
 }}

\item{stationData}{default is \code{NULL}.  This is the path to a csv file containing the all of 
the strata for which stations should be generated.  In addition to the strata, this file must 
also include the fields  \code{"PRIMARY"}, \code{"SECONDARY"}, and \code{"ALTERNATE"}.  These 
three fields should contain integers corresponding with how many of that type of station should 
be generated.  If no stations of a particular type should be generated, a value of NA should be 
present. Strata where \code{PRIMARY} is set to NA will not have any stations generated. Below is 
an example of how this file might look:
> head(stationFile)
  STRATUM PRIMARY ALTERNATE SECONDARY
      5Z1       6         3         4
      5Z2      11         6        NA
      5Z3       5         2         2
      5Z4       4         2         2}

\item{stationData_StratField}{default is \code{"STRATUM"}.  This is just the name of the field 
that contains the strata.  For the example above, this would be "STRATUM".}

\item{strataShp}{default is \code{NULL}. This should point to the ".shp" file of a shapefile for 
the strata to use. This shapefile must have a field that has values identical to those found in 
the \code{"stationData_StratField"} within the \code{"stationData"} file.}

\item{strataShp_StratField}{default is \code{NULL}.  This is the name of the field within 
\code{strataShp} that contains the identifier for the strata.  Continuing with the example above,
this field would contain values including "5Z1", "5Z2", "5Z3" and "5Z4".}

\item{avoidShp}{default is \code{NULL}. This should point to the ".shp" file of a shapefile 
containing polygons of the areas where stations should not be located.   For example, one might 
populate a file with areas known to contain unexploded ordinance, or areas where bottom contact 
is forbidden.  No stations will be generated where polygons exist in this file.}

\item{localCRS}{default is \code{2961}.  In order to create sampling stations, the function 
reprojects any spatial data to a locally appropriate projection.  The default value is UTM Zone 
20N, and is appropriate only for Maritimes data.  Any valid CRS can be entered here, and should 
be appropriate for your data.}

\item{minDistNM}{default is \code{4}. This is the minimum required distance between the sets.  By
default, sets will be no closer than 4 nautical miles from each other.}

\item{addNafoInfo}{default is \code{TRUE}.  Should NAFO area information be part of the output data?}

\item{oceansAreas}{default is \code{NULL}. This is a shapefile that contains polygons that have 
information that should be associated with the selected stations.  The idea is that this can be 
used to  associate important information with points that fall within that area. 
that fall}

\item{tryXTimes}{default is \code{100}  By default, the script will make this many attempts to 
fit the requested number of stations into each strata.}
}
\description{
This function randomly selects stations within particular strata.  A specified 
number of stations of 3 different categories can be selected, and the minimum distance required 
between the stations can be specified.  Additionally, station selection can avoid particular 
areas through the inclusion of an additional shapefile.
}
\examples{
\dontrun{
NonMaritimesExample <-setSelect(minDistNM = 4, 
                                stationData = "mySurveyStrata_requirements.csv", 
                                stationData_StratField = "STRATUM", 
                                strataShp = "myStrata.shp", 
                                strataShp_StratField = "StrataID",
                                avoidShp = "areasWeCantTrawl.shp", 
                                localCRS = 2961, 
                                MaritimesSurvey = NULL)

MaritimesExample   <- setSelect(MaritimesSurvey = "GEORGES_5Z",
                                avoidShp = "Maritimes2022_AvoidAreas.shp")
                       }
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{surveyTools}
