% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{setTweaks}
\alias{setTweaks}
\title{setTweaks}
\usage{
setTweaks(x = NULL)
}
\arguments{
\item{x}{default is \code{NULL}.  Set data frame}
}
\value{
a  Set data frame
}
\description{
This function will perform tweaks to data coming from ANdes before 
it is imported to Oracle.
}
\seealso{
Other general_use: 
\code{\link{addSizeClassToCatch}()},
\code{\link{basketTweaks}()},
\code{\link{catchTweaks}()},
\code{\link{formatForOracle}()},
\code{\link{loadData}()},
\code{\link{lv1Tweaks}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{specimenTweaks}()},
\code{\link{toOracle}()},
\code{\link{universalTweaks}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
