% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{SPRING_4X}
\alias{SPRING_4X}
\title{SPRING_4X Metadata}
\format{
data.frame
\describe{
\item{STRATUM}{survey stratum code}
\item{PRIMARY}{integer indicating required number of primary stations}
\item{ALTERNATE}{integer indicating required number of alternate stations}
\item{SECONDARY}{integer indicating required number of secondary stations}
}
}
\usage{
SPRING_4X
}
\description{
This data.frame includes all of the strata sampled during modern Spring 4X surveys
and the required number of primary, secondary and alternate stations for each.
}
\keyword{datasets}
