% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\docType{data}
\name{SUMMER_4VWX}
\alias{SUMMER_4VWX}
\title{SUMMER_4VWX Metadata}
\format{
data.frame
\describe{
\item{STRATUM}{survey stratum code}
\item{PRIMARY}{integer indicating required number of primary stations}
\item{ALTERNATE}{integer indicating required number of alternate stations}
\item{SECONDARY}{integer indicating required number of secondary stations}
}
}
\usage{
SUMMER_4VWX
}
\description{
This data.frame includes all of the strata sampled during modern Summer 4VWX surveys
and the required number of primary, secondary and alternate stations for each.
}
\keyword{datasets}
