% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertUtilities.R
\name{TF2YN}
\alias{TF2YN}
\title{convertHOWOBT}
\usage{
TF2YN(x)
}
\arguments{
\item{x}{default is \code{NULL}. This is the string to be processed.}

\item{field}{default is \code{NULL}. This is the field that should be converted.}
}
\value{
converted FORCE codes

vector
}
\description{
This function returns converted obtained_code codes

This function converts TRUE/FALSE values to Y/N
}
\seealso{
Other internal: 
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()}

Other internal: 
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}

Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
