% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toOracle.R
\name{toOracle}
\alias{toOracle}
\title{toOracle}
\usage{
toOracle(
  cxnObj = NULL,
  source_df = NULL,
  target_schema = "groundfish",
  target_table = NULL,
  createReplaceTarget = FALSE
)
}
\arguments{
\item{cxnObj}{default is \code{NULL}.  This is a connection object from 
Mar.utils::make_oracle_cxn(). The account specified by this needs permission to write tables in 
the \code{target_schema}.}

\item{source_df}{default is \code{NULL}.  This is a data frame containing the data that should be 
loaded to Oracle.  It should correspond with the structure used by one of ESE_MISSIONS, ESE_SETS, 
ESE_CATCHES, ESE_BASKETS, ESE_SPECIMENS, ESE_LV1_OBSERVATIONS}

\item{target_schema}{default is \code{"groundfish"}.  This is the name of the Oracle schema where the 
data stored in\code{source_df} should be loaded.}

\item{target_table}{default is \code{NULL}.  This is the name of the specific Oracle table within
\code{target_schema} where the data should be loaded}

\item{createReplaceTarget}{default is \code{FALSE}.  By default, new data is appended to tables
that already exist within the specified \code{target_schema}.  Setting this to \code{TRUE} will 
result in the creation of new tables or the replacement of existing tables within 
\code{target_schema}.  Use with caution.}
}
\value{
nothing, but data gets loaded to the selected location in Oracle.
}
\description{
This function loads any of 6 known data frame objects into Oracle, ensuring that the 
records of the new objects will not replace any existing records.
}
\seealso{
Other general_use: 
\code{\link{addSizeClassToCatch}()},
\code{\link{basketTweaks}()},
\code{\link{catchTweaks}()},
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{setTweaks}()},
\code{\link{specimenTweaks}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
