% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmogrify.R
\name{transmogrifyLV1_OBS}
\alias{transmogrifyLV1_OBS}
\title{transmogrifyLV1_OBS}
\usage{
transmogrifyLV1_OBS(df = NULL)
}
\arguments{
\item{df}{default is \code{NULL}.  This is the data frame containing the usable fields from the 
specimen_data. Note that both ESE_SPECIMENS and ESE_LV1_OBSERVATIONS are derived from the andes 
specimen_data.}
}
\description{
This function takes the fields from the andes specimen_data, and does the various 
processing necessary to make them equivalent to the formats used in the ESE_LV1_OBSERVATIONS table
}
\seealso{
Other internal: 
\code{\link{applySubsampling}()},
\code{\link{charToBinary}()},
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{keepFieldsBaskets}()},
\code{\link{keepFieldsCatches}()},
\code{\link{keepFieldsMissions}()},
\code{\link{keepFieldsSets}()},
\code{\link{keepFieldsSpecimens}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()},
\code{\link{transmogrifyBaskets}()},
\code{\link{transmogrifyCatches}()},
\code{\link{transmogrifyMissions}()},
\code{\link{transmogrifySets}()},
\code{\link{transmogrifySpecimens}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
