% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{tweakBaskets}
\alias{tweakBaskets}
\title{tweakBaskets}
\usage{
tweakBaskets(x = NULL)
}
\arguments{
\item{x}{default is \code{NULL}.  This is the ESE_BASKET data frame}
}
\value{
a data frame for loading into
}
\description{
This function will perform tweaks to data coming from andes before 
it is imported to Oracle.
}
\seealso{
Other general_use: 
\code{\link{colTypeConverter}()},
\code{\link{loadData}()},
\code{\link{loadESEData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{tweakBasketsPostProcessing}()},
\code{\link{tweakCatches}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimensRaw}()},
\code{\link{tweakUniversal}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
