% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{tweakBasketsPostProcessing}
\alias{tweakBasketsPostProcessing}
\title{tweakBasketsPostProcessing}
\usage{
tweakBasketsPostProcessing(basket = NULL, lv1 = NULL)
}
\arguments{
\item{basket}{default is \code{NULL}.}

\item{lv1}{default is \code{NULL}.}
}
\value{
an updated version of ESE_BASKETS
}
\description{
This function can do final changes to the basket data once all of the various basket/
catch handling and re-allocation of mixed catches is complete.  This processing can include using 
measured weights to change basket weights, so both basket and lv1 objects are necessary as 
parameters.
}
\seealso{
Other general_use: 
\code{\link{colTypeConverter}()},
\code{\link{loadData}()},
\code{\link{loadESEData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimensRaw}()},
\code{\link{tweakUniversal}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
