% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{tweakCatches}
\alias{tweakCatches}
\title{tweakCatches}
\usage{
tweakCatches(x = NULL)
}
\arguments{
\item{x}{default is \code{NULL}.   This is the ESE_CATCHES data frame}
}
\value{
a data frame for loading into ESE_CATCHES
}
\description{
This function will perform tweaks to data coming from andes before 
it is imported to Oracle.
}
\seealso{
Other general_use: 
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{replaceESEData}()},
\code{\link{setExperimentType}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimens}()},
\code{\link{tweakUniversal}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
