% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{tweakLv1}
\alias{tweakLv1}
\title{tweakLv1}
\usage{
tweakLv1(x = NULL)
}
\arguments{
\item{x}{default is \code{NULL}.  This is the ESE_LV1_OBSERVATIONS data frame}
}
\value{
an updated version of ESE_LV1_OBSERVATIONS
}
\description{
This function will perform tweaks to the lv1_observations data coming from andes before 
it is imported to Oracle.
}
\seealso{
Other general_use: 
\code{\link{colTypeConverter}()},
\code{\link{loadData}()},
\code{\link{loadESEData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{tweakBasketsPostProcessing}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimensRaw}()},
\code{\link{tweakUniversal}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
