% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{tweakSets}
\alias{tweakSets}
\title{tweakSets}
\usage{
tweakSets(x = NULL)
}
\arguments{
\item{x}{default is \code{NULL}.  This is the ESE_SETS data frame.}
}
\value{
a data frame for loading into ESE_SETS
}
\description{
This function will perform tweaks to data coming from andes before 
it is imported to Oracle.
}
\seealso{
Other general_use: 
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{replaceESEData}()},
\code{\link{setExperimentType}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakSpecimensRaw}()},
\code{\link{tweakUniversal}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
