% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{tweakSpecimens}
\alias{tweakSpecimens}
\title{tweakSpecimens}
\usage{
tweakSpecimens(x = NULL)
}
\arguments{
\item{x}{default is \code{NULL}. This is the ESE_BASKETS data frame}
}
\value{
a data frame for loading into ESE_BASKETS
}
\description{
This function will perform tweaks to data coming from andes before 
it is imported to Oracle.
}
\seealso{
Other general_use: 
\code{\link{addSizeClassToCatch}()},
\code{\link{formatForOracle}()},
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{toOracle}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakUniversal}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
