% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{tweakSpecimensRaw}
\alias{tweakSpecimensRaw}
\title{tweakSpecimensRaw}
\usage{
tweakSpecimensRaw(x = NULL)
}
\arguments{
\item{x}{default is \code{NULL}. This is the raw "specimen_data" data frame}
}
\value{
a data frame for loading into ESE_BASKETS
}
\description{
This function will perform tweaks to data coming from andes before 
it is imported to Oracle.  specimen_data gets converted into both ESE_SPECIMENS and 
ESE_LV1_OBSERVATIONS, so many tweaks should be performed here to ensure that changes are 
reflected in both output tables.  Also, changing here will ensure sequential LV1_OBSERVATION_ID 
values
}
\seealso{
Other general_use: 
\code{\link{colTypeConverter}()},
\code{\link{loadData}()},
\code{\link{loadESEData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{tweakBasketsPostProcessing}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakUniversal}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
