% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{universalTweaks}
\alias{universalTweaks}
\title{universalTweaks}
\usage{
universalTweaks(x = NULL)
}
\arguments{
\item{x}{default is \code{NULL}.  This is a list of all of the objects from the csv files}
}
\value{
list
}
\description{
This function will perform tweaks on data from Andes that exists in multiple 
exported csv files
}
\seealso{
Other general_use: 
\code{\link{addSizeClassToCatch}()},
\code{\link{basketTweaks}()},
\code{\link{catchTweaks}()},
\code{\link{formatForOracle}()},
\code{\link{loadData}()},
\code{\link{lv1Tweaks}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{setTweaks}()},
\code{\link{specimenTweaks}()},
\code{\link{toOracle}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
