% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{addSizeClassToCatch}
\alias{addSizeClassToCatch}
\title{addSizeClassToCatch}
\usage{
addSizeClassToCatch(basket, catch, quiet = FALSE)
}
\arguments{
\item{basket}{A data frame containing the basket information from Andes}

\item{catch}{A data frame containing the catch information from Andes}

\item{quiet}{default is \code{FALSE} Determines whether or not the script should output 
informational messages}
}
\value{
the merged catch card
}
\description{
This function fills in the catch with the missing size class 2 entries.   
Andes does not provide a way to add Size class directly, Size classes are entered via the basket form.
}
\seealso{
Other general_use: 
\code{\link{basketTweaks}()},
\code{\link{catchTweaks}()},
\code{\link{eseExtractor}()},
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setExperimentType}()},
\code{\link{setTweaks}()},
\code{\link{specimenTweaks}()},
\code{\link{toOracle}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
