% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanEse}
\alias{cleanEse}
\title{cleanEse}
\usage{
cleanEse()
}
\value{
nothing - just remove data frames from environment
}
\description{
This function removes any ESE objects from the R environment
}
\seealso{
Other internal: 
\code{\link{cleanfields}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertLengthUnits}()},
\code{\link{convertSex}()},
\code{\link{convertStrata}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{convertYesNo}()},
\code{\link{getEseTables}()},
\code{\link{getKeyFields}()},
\code{\link{getTblKey}()},
\code{\link{get_value}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
