% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cleanEse}
\alias{cleanEse}
\title{cleanEse}
\usage{
cleanEse()
}
\value{
nothing - just remove data frames from environment
}
\description{
This function removes any ESE objects from the R environment
}
\seealso{
Other internal: 
\code{\link{charToBinary}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{keepFieldsBaskets}()},
\code{\link{keepFieldsCatches}()},
\code{\link{keepFieldsMissions}()},
\code{\link{keepFieldsSets}()},
\code{\link{keepFieldsSpecimens}()},
\code{\link{meters2Fathoms}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()},
\code{\link{redistributeMixedCatch2}()},
\code{\link{transmogrifyBaskets}()},
\code{\link{transmogrifyCatches}()},
\code{\link{transmogrifyLV1_OBS}()},
\code{\link{transmogrifyMissions}()},
\code{\link{transmogrifySets}()},
\code{\link{transmogrifySpecimens}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
