% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.utilities.R
\name{convertSex}
\alias{convertSex}
\title{convertSex}
\usage{
convertSex(x)
}
\arguments{
\item{x}{default is \code{NULL}. This is the string to be processed.}
}
\value{
converted sex codes
}
\description{
This function returns converted sex codes
}
\seealso{
Other internal: 
\code{\link{cleanEse}()},
\code{\link{cleanfields}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertLengthUnits}()},
\code{\link{convertStrata}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{convertYesNo}()},
\code{\link{getEseTables}()},
\code{\link{getKeyFields}()},
\code{\link{getTblKey}()},
\code{\link{get_value}()},
\code{\link{reFormatSpecimen}()}
}
\author{
Pablo Vergara, \email{pablo.vergara@dfo-mpo.gc.ca}
}
\concept{internal}
