% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.utilities.R
\name{get_value}
\alias{get_value}
\title{get_value}
\usage{
get_value(myKey, mylookupvector)
}
\value{
lookup value
}
\description{
This function returns the  lookup value in a vector
}
\seealso{
Other internal: 
\code{\link{base}},
\code{\link{cleanEse}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertLengthUnits}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{convertYesNo}()},
\code{\link{createInsertIntoObsFields}()},
\code{\link{createInsertIntoObsGrp}()},
\code{\link{createInsertIntoSampReq}()},
\code{\link{getEseTables}()},
\code{\link{getKeyFields}()},
\code{\link{getTblKey}()}
}
\author{
Pablo Vergara, \email{pablo.vergara@dfo-mpo.gc.ca}
}
\concept{internal}
