% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getKeyFields}
\alias{getKeyFields}
\title{getKeyFields}
\usage{
getKeyFields(table = NULL)
}
\arguments{
\item{table}{default is \code{NULL}.  This is the name of the table for which the key fields are 
desired}
}
\value{
nothing - just loads data to environment
}
\description{
This function returns the key fields for any of the ESE tables
}
\seealso{
Other internal: 
\code{\link{base}},
\code{\link{cleanEse}()},
\code{\link{convertFieldType}()},
\code{\link{convertLengthType}()},
\code{\link{convertLengthUnits}()},
\code{\link{convertWeighType}()},
\code{\link{convertWeightUnits}()},
\code{\link{convertYesNo}()},
\code{\link{createInsertIntoObsFields}()},
\code{\link{createInsertIntoObsGrp}()},
\code{\link{createInsertIntoSampReq}()},
\code{\link{getEseTables}()},
\code{\link{getTblKey}()},
\code{\link{get_value}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
