% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchAndesToESE.R
\name{matchAndesToESE}
\alias{matchAndesToESE}
\title{matchAndesToESE}
\usage{
matchAndesToESE(dataPath = NULL)
}
\arguments{
\item{dataPath}{default is \code{NULL}. This is the location of the csv files exported from andes}
}
\value{
a list with all objects needed to load to Oracle
}
\description{
This function matches the new CSV dump from Andes to the old ESE table schema so that data can
be loaded into Oracle.
}
\seealso{
Other general_use: 
\code{\link{loadData}()},
\code{\link{replaceESEData}()},
\code{\link{setExperimentType}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakLv1}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimens}()},
\code{\link{tweakUniversal}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
