% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchAndesToESE.R
\name{matchAndesToESE}
\alias{matchAndesToESE}
\title{matchAndesToESE}
\usage{
matchAndesToESE(dataPath = NULL, quiet = FALSE)
}
\arguments{
\item{dataPath}{default is \code{NULL}.  This is the location of the csv files exported from andes}

\item{quiet}{default is \code{FALSE} Determines whether or not the script should output 
informational messages}
}
\value{
a list with all objects needed to load to Oracle
}
\description{
This function matches the new CSV dump from Andes to the old ESE table schema so that data can
be loaded into Oracle.
}
\seealso{
Other general_use: 
\code{\link{addSizeClassToCatch}()},
\code{\link{basketTweaks}()},
\code{\link{catchTweaks}()},
\code{\link{loadData}()},
\code{\link{setExperimentType}()},
\code{\link{setTweaks}()},
\code{\link{specimenTweaks}()},
\code{\link{toOracle}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
