% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{meters2Fathoms}
\alias{meters2Fathoms}
\title{meters2Fathoms}
\usage{
meters2Fathoms(field = NULL)
}
\arguments{
\item{field}{default is \code{NULL}. This is the field that should be converted.}
}
\value{
vector
}
\description{
This function converts a vector of values from meter to fathoms
}
\seealso{
Other internal: 
\code{\link{applySubsampling}()},
\code{\link{charToBinary}()},
\code{\link{cleanEse}()},
\code{\link{cleanStrata}()},
\code{\link{cleanfields}()},
\code{\link{convertFORCE}()},
\code{\link{convertHOWOBT}()},
\code{\link{getEseTables}()},
\code{\link{get_value}()},
\code{\link{keepFieldsBaskets}()},
\code{\link{keepFieldsCatches}()},
\code{\link{keepFieldsMissions}()},
\code{\link{keepFieldsSets}()},
\code{\link{keepFieldsSpecimens}()},
\code{\link{populate_DATA_DESC}()},
\code{\link{reFormatSpecimen}()},
\code{\link{transmogrifyBaskets}()},
\code{\link{transmogrifyCatches}()},
\code{\link{transmogrifyLV1_OBS}()},
\code{\link{transmogrifyMissions}()},
\code{\link{transmogrifySets}()},
\code{\link{transmogrifySpecimens}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
