% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{reFormatSpecimen}
\alias{reFormatSpecimen}
\title{reFormatSpecimen}
\usage{
reFormatSpecimen(x = NULL, quiet = FALSE)
}
\arguments{
\item{x}{default is \code{NULL}.  This is a field in a data frame}

\item{quiet}{default is \code{FALSE} Determines whether or not the script should output 
informational messages}
}
\description{
This function takes the tempSpecimen object, and populates the DATA_DEC 
appropriately depending on the values found
}
\seealso{
Other internal: 
\code{\link{cleanEse}()},
\code{\link{cleanfields}()},
\code{\link{getEseTables}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{internal}
