% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setExperimentType.R
\name{setExperimentType}
\alias{setExperimentType}
\title{setExperimentType}
\usage{
setExperimentType(x, quiet = FALSE)
}
\arguments{
\item{x}{set object from ANdes raw data}

\item{quiet}{default is \code{FALSE} Determines whether or not the script should output 
informational messages}
}
\value{
The experiment code
}
\description{
This function sets the proper experiment type to match existing data structures in Maritime and Gulf region
. Based on a mix of experiment typer and set results
}
\seealso{
Other general_use: 
\code{\link{addSizeClassToCatch}()},
\code{\link{basketTweaks}()},
\code{\link{catchTweaks}()},
\code{\link{eseExtractor}()},
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{setTweaks}()},
\code{\link{specimenTweaks}()},
\code{\link{toOracle}()}
}
\author{
Pablo Vergara, \email{Pablo.Vergara@dfo-mpo.gc.ca}
}
\concept{general_use}
