% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweaks.R
\name{tweakUniversal}
\alias{tweakUniversal}
\title{tweakUniversal}
\usage{
tweakUniversal(x = NULL, mission = NULL)
}
\arguments{
\item{x}{default is \code{NULL}.  This is a list of all of the objects from the csv files}

\item{mission}{default is \code{NULL}. This is the cleaned mission id (e.g. CAR2022102)}
}
\value{
list
}
\description{
This function will perform tweaks on data from Andes that exists in multiple 
exported csv files
}
\seealso{
Other general_use: 
\code{\link{loadData}()},
\code{\link{matchAndesToESE}()},
\code{\link{replaceESEData}()},
\code{\link{setExperimentType}()},
\code{\link{tweakBaskets}()},
\code{\link{tweakCatches}()},
\code{\link{tweakSets}()},
\code{\link{tweakSpecimensRaw}()}
}
\author{
Mike McMahon, \email{Mike.McMahon@dfo-mpo.gc.ca}
}
\concept{general_use}
